#6 Installation and configuration of the Admin panel
Configuration of the admin part with AdminLTE. Database data is correct. Models with tables and controllers are configured on base level. Correct setting of validator (validation of data). Vue components for display and create records in "database table".  Installed vform for validation of data for create-query. Customization of styles thanks to fontawesome.